refact: exitChatRoom -> chat.ts
#42
