+++ src/pages/auth/chat/chatRoom/Exit.tsx
@@ -1,26 +1,10 @@
 import { useNavigate } from "react-router-dom";
-import { getSocket } from "socket/socket";
-
-interface dataType {
-  status: 'error' | 'approved'
-  detail: string
-}
+import { exitChatRoom } from "ws/chat";
 
 export default function Exit(props: { room: number }) {
   const navigate = useNavigate();
   function exitHandler() {
-    const socket = getSocket();
-    if (socket) {
-      socket.emit('exitChatRoom', {
-        "roomId": props.room
-      });
-      socket.on('exitChatRoomResult', (data) => {
-        const res: dataType = data;
-        if (res.status === 'approved')
-          navigate('/chat/list');
-        console.log(data);
-      })
-    }
+    exitChatRoom(props.room, navigate);
   }
   return (
     <button onClick={exitHandler}>나가기</button>
